Split some `org.embulk.config` classes into `embulk-api`
